backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Data Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Date/Time
python-dateutil>=2.8.2

# SMS Notifications (Optional)
twilio>=8.11.1

# HTTP Client
httpx>=0.26.0
requests>=2.31.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3

# CORS
python-multipart>=0.0.6
phonenumbers>=8.13.29
bcrypt>=4.1.2

# Google OAuth
google-auth>=2.28.0
PyJWT>=2.8.0

# Push Notifications
exponent-server-sdk>=2.2.0
